WIP: T03 -- Split tests/unit/test_autodetect_apps.py along its four existing class boundaries
