+++ tests/unit/test_autodetect_apps.py
@@ -1,16 +1,13 @@
 """Tests for auto-detect apps functionality."""
 
-import logging
 from pathlib import Path
 from textwrap import dedent
-from unittest.mock import patch
 
 import pytest
 
 from hassette import context
 from hassette.config.config import HassetteConfig
 from hassette.config.defaults import AUTODETECT_EXCLUDE_DIRS_DEFAULT
-from hassette.test_utils.config import TEST_TOKEN, make_test_config
 from hassette.utils.app_utils import autodetect_apps
 
 
@@ -381,544 +378,3 @@ class TestAutoDetectApps:
         assert "apps.mixed_classes.ActualApp" in result, (
             "Expected to find 'apps.mixed_classes.ActualApp' in detected apps"
         )
-
-
-class TestValidateApps:
-    """Test the validate_apps function."""
-
-    def make_config(
-        self,
-        tmp_path: Path,
-        *,
-        autodetect: bool = False,
-        directory: Path | None = None,
-        apps: dict | None = None,
-    ) -> HassetteConfig:
-        apps_cfg: dict = {"autodetect": autodetect}
-        if directory is not None:
-            apps_cfg["directory"] = directory
-        if apps is not None:
-            apps_cfg["apps"] = apps
-        return make_test_config(data_dir=tmp_path, apps=apps_cfg)
-
-    def test_validate_apps_sets_app_dir(self, tmp_path: Path):
-        """Test that validate_apps sets app_dir for apps that don't have it."""
-        app_dir = tmp_path / "test_apps"
-        app_dir.mkdir(parents=True, exist_ok=True)
-        config = self.make_config(
-            tmp_path,
-            directory=app_dir,
-            apps={"my_app": {"filename": "my_app.py", "class_name": "MyApp"}},
-        )
-
-        with context.use_hassette_config(config):
-            config.set_validated_app_manifests()
-        results = config.apps.manifests
-
-        assert results["my_app"].app_dir == app_dir, (
-            f"Expected app_dir to be {app_dir}, got {results['my_app'].app_dir}"
-        )
-        assert results["my_app"].app_key == "my_app", (
-            f"Expected app_key to be 'my_app', got {results['my_app'].app_key}"
-        )
-
-    def test_validate_apps_preserves_existing_app_dir(self, tmp_path: Path):
-        """Test that validate_apps preserves existing app_dir values."""
-        app_dir = tmp_path / "test_apps"
-        app_dir.mkdir(parents=True, exist_ok=True)
-        custom_dir = Path("/custom/location")
-
-        config = self.make_config(
-            tmp_path,
-            directory=app_dir,
-            apps={"my_app": {"filename": "my_app.py", "class_name": "MyApp", "app_dir": custom_dir}},
-        )
-
-        with context.use_hassette_config(config):
-            config.set_validated_app_manifests()
-        results = config.apps.manifests
-
-        assert results["my_app"].app_dir == custom_dir, (
-            f"Expected app_dir to be {custom_dir}, got {results['my_app'].app_dir}"
-        )
-
-    def test_validate_apps_removes_invalid_apps(self, tmp_path: Path):
-        """Test that validate_apps removes apps missing required keys."""
-        app_dir = tmp_path / "test_apps"
-        app_dir.mkdir(parents=True, exist_ok=True)
-
-        config = self.make_config(
-            tmp_path,
-            directory=app_dir,
-            apps={
-                "valid_app": {
-                    "filename": "valid.py",
-                    "class_name": "ValidApp",
-                },
-                "missing_filename": {
-                    "class_name": "MissingFilename",
-                },
-                "missing_class_name": {
-                    "filename": "missing_class.py",
-                },
-                "missing_both": {
-                    "some_config": "value",
-                },
-            },
-        )
-
-        with context.use_hassette_config(config):
-            config.set_validated_app_manifests()
-        result = config.apps.manifests
-
-        # Only valid_app should remain - this is the important functional test
-        assert len(result) == 1, f"Expected 1 valid app, got {len(result)}"
-        assert "valid_app" in result, "Expected to find 'valid_app' in detected apps"
-        assert "missing_filename" not in result, "Did not expect to find 'missing_filename' in detected apps"
-        assert "missing_class_name" not in result, "Did not expect to find 'missing_class_name' in detected apps"
-        assert "missing_both" not in result, "Did not expect to find 'missing_both' in detected apps"
-
-        # The valid app should have the app_dir and app_key set
-        assert result["valid_app"].app_dir == app_dir, (
-            f"Expected app_dir to be {app_dir}, got {result['valid_app'].app_dir}"
-        )
-        assert result["valid_app"].app_key == "valid_app", (
-            f"Expected app_key to be 'valid_app', got {result['valid_app'].app_key}"
-        )
-
-    @patch("hassette.config.config.autodetect_apps")
-    def test_validate_apps_calls_autodetect(self, mock_autodetect, tmp_path: Path):
-        """Test that validate_apps calls autodetect_app_manifests when autodetect=True."""
-        app_dir = tmp_path / "test_apps"
-        app_dir.mkdir(parents=True, exist_ok=True)
-
-        # Mock the auto-detection to return a detected app
-        mock_app_dict = {
-            "filename": "auto.py",
-            "class_name": "AutoApp",
-            "app_dir": app_dir,
-            "app_key": "auto.AutoApp",
-            "enabled": True,
-            "full_path": app_dir / "auto.py",
-        }
-        mock_autodetect.return_value = {"auto.AutoApp": mock_app_dict}
-
-        config = self.make_config(
-            tmp_path,
-            autodetect=True,
-            directory=app_dir,
-            apps={
-                "manual_app": {
-                    "filename": "manual.py",
-                    "class_name": "ManualApp",
-                }
-            },
-        )
-
-        with context.use_hassette_config(config):
-            config.set_validated_app_manifests()
-        result = config.apps.manifests
-
-        # Should have both manual and auto-detected apps
-        assert len(result) == 2, f"Expected 2 apps, got {len(result)}"
-        assert "manual_app" in result, "Expected to find 'manual_app' in detected apps"
-        assert "auto.AutoApp" in result, "Expected to find 'auto.AutoApp' in detected apps"
-
-        # Check that autodetect_app_manifests was called with correct parameters
-        mock_autodetect.assert_called_once()
-        args, _ = mock_autodetect.call_args
-        assert args[0] == app_dir, f"Expected app_dir to be {app_dir}, got {args[0]}"
-        # Should include the path of the manual app in known_paths
-        known_paths = args[1]
-        expected_path = (app_dir / "manual.py").resolve()
-        assert expected_path in known_paths, f"Expected known_paths to include {expected_path}, got {known_paths}"
-
-    @patch("hassette.config.config.autodetect_apps")
-    def test_validate_apps_skips_conflicting_autodetected(self, mock_autodetect, tmp_path: Path):
-        """Test that validate_apps skips auto-detected apps that conflict with manual ones."""
-        app_dir = tmp_path / "test_apps"
-        app_dir.mkdir(parents=True, exist_ok=True)
-
-        # Mock auto-detection to return an app with the same key
-        mock_app_dict = {
-            "filename": "my_app.py",
-            "class_name": "MyApp",
-            "app_dir": app_dir,
-            "app_key": "my_app",
-            "enabled": True,
-        }
-        mock_autodetect.return_value = {"my_app": mock_app_dict}
-
-        config = self.make_config(
-            tmp_path,
-            autodetect=True,
-            directory=app_dir,
-            apps={
-                "my_app": {
-                    "filename": "my_app.py",
-                    "class_name": "MyApp",
-                }
-            },
-        )
-
-        with context.use_hassette_config(config):
-            config.set_validated_app_manifests()
-        result = config.apps.manifests
-
-        # Should only have the manual app, not the auto-detected one
-        assert len(result) == 1, f"Expected 1 app, got {len(result)}"
-        assert "my_app" in result, "Expected to find 'my_app' in detected apps"
-        # Should be the original manual config, not the auto-detected one
-        assert result["my_app"].filename == "my_app.py", (
-            f"Expected filename to be 'my_app.py', got {result['my_app'].filename}"
-        )
-        assert result["my_app"].class_name == "MyApp", (
-            f"Expected class_name to be 'MyApp', got {result['my_app'].class_name}"
-        )
-
-        # The behavior is that auto-detected apps with conflicting keys are simply not added
-        # Let's verify that autodetect_app_manifests was called and the result is correct
-        mock_autodetect.assert_called_once()
-        # The important thing is that the result only contains the manual app
-
-    def test_validate_apps_skips_autodetect_when_disabled(self, tmp_path: Path):
-        """Test that validate_apps skips auto-detection when autodetect=False."""
-        app_dir = tmp_path / "test_apps"
-        app_dir.mkdir(parents=True, exist_ok=True)
-
-        config = self.make_config(
-            tmp_path,
-            directory=app_dir,
-            apps={
-                "manual_app": {
-                    "filename": "manual.py",
-                    "class_name": "ManualApp",
-                }
-            },
-        )
-
-        with patch("hassette.config.config.autodetect_apps") as mock_autodetect:
-            with context.use_hassette_config(config):
-                config.set_validated_app_manifests()
-            result = config.apps.manifests
-
-            # Should not call autodetect_apps since autodetect=False
-            mock_autodetect.assert_not_called()
-
-            # Should only have the manual app
-            assert len(result) == 1, f"Expected 1 app, got {len(result)}"
-            assert "manual_app" in result, "Expected to find 'manual_app' in detected apps"
-
-    def test_validate_apps_warns_on_cache_key_collision(self, tmp_path: Path, caplog: pytest.LogCaptureFixture) -> None:
-        """Two apps with different app_key but the same explicit cache_key log a WARNING."""
-        # Some other test in this session may have left the "hassette" logger's propagate flag
-        # set to False (e.g. via enable_basic_logging()); caplog relies on propagation to the
-        # root logger, so restore it here. See src/hassette/test_utils/harness.py:337-340 for
-        # the same workaround applied elsewhere.
-        logging.getLogger("hassette").propagate = True
-        app_dir = tmp_path / "test_apps"
-        app_dir.mkdir(parents=True, exist_ok=True)
-
-        config = self.make_config(
-            tmp_path,
-            directory=app_dir,
-            apps={
-                "app_one": {
-                    "filename": "app_one.py",
-                    "class_name": "AppOne",
-                    "cache_key": "shared-key",
-                },
-                "app_two": {
-                    "filename": "app_two.py",
-                    "class_name": "AppTwo",
-                    "cache_key": "shared-key",
-                },
-            },
-        )
-
-        with context.use_hassette_config(config), caplog.at_level("WARNING", logger="hassette.config.config"):
-            config.set_validated_app_manifests()
-
-        assert any("shared-key" in record.message for record in caplog.records), (
-            f"Expected a WARNING mentioning the colliding cache_key, got: {[r.message for r in caplog.records]}"
-        )
-
-    def test_validate_apps_no_warning_when_cache_keys_unique(
-        self, tmp_path: Path, caplog: pytest.LogCaptureFixture
-    ) -> None:
-        """Apps with distinct (or default) cache_keys produce no collision warning."""
-        app_dir = tmp_path / "test_apps"
-        app_dir.mkdir(parents=True, exist_ok=True)
-
-        config = self.make_config(
-            tmp_path,
-            directory=app_dir,
-            apps={
-                "app_one": {"filename": "app_one.py", "class_name": "AppOne"},
-                "app_two": {"filename": "app_two.py", "class_name": "AppTwo"},
-            },
-        )
-
-        with context.use_hassette_config(config), caplog.at_level("WARNING", logger="hassette.config.config"):
-            config.set_validated_app_manifests()
-
-        assert not any("cache_key" in record.message for record in caplog.records), (
-            f"Expected no cache_key collision warning, got: {[r.message for r in caplog.records]}"
-        )
-
-    def test_validate_apps_warns_on_multi_instance_default_key_collision(
-        self, tmp_path: Path, caplog: pytest.LogCaptureFixture
-    ) -> None:
-        """A multi-instance app's default `{app_key}/{idx}` key can collide with another
-        app's explicit cache_key — the collision check must expand multi-instance app_config
-        lists to each instance's resolved key, not just check the manifest as a whole.
-        """
-        # See comment in test_validate_apps_warns_on_cache_key_collision above.
-        logging.getLogger("hassette").propagate = True
-        app_dir = tmp_path / "test_apps"
-        app_dir.mkdir(parents=True, exist_ok=True)
-
-        config = self.make_config(
-            tmp_path,
-            directory=app_dir,
-            apps={
-                "weather": {
-                    "filename": "weather.py",
-                    "class_name": "Weather",
-                    "config": [{"city": "nyc"}, {"city": "sf"}],
-                },
-                "other_app": {
-                    "filename": "other_app.py",
-                    "class_name": "OtherApp",
-                    "cache_key": "weather/1",
-                },
-            },
-        )
-
-        with context.use_hassette_config(config), caplog.at_level("WARNING", logger="hassette.config.config"):
-            config.set_validated_app_manifests()
-
-        assert any("weather/1" in record.message for record in caplog.records), (
-            f"Expected a WARNING mentioning the colliding cache_key 'weather/1', "
-            f"got: {[r.message for r in caplog.records]}"
-        )
-
-
-class TestAutoDetectIntegration:
-    """Integration tests for auto-detect functionality with HassetteConfig."""
-
-    @pytest.fixture(autouse=True)
-    def setup(self, test_config: HassetteConfig):
-        with context.use_hassette_config(test_config):
-            yield
-
-    def test_hassette_config_autodetect_enabled_by_default(self, tmp_path: Path):
-        """Test that autodetect_apps is enabled by default in HassetteConfig."""
-        # Create a temporary app directory with an app
-        app_dir = tmp_path / "apps"
-        app_dir.mkdir()
-
-        app_file = app_dir / "test_app.py"
-        app_file.write_text(
-            dedent("""
-            from hassette import App, AppConfig
-
-            class TestApp(App[AppConfig]): ...
-        """)
-        )
-
-        # Create config with the temp app directory
-        config = HassetteConfig(token=TEST_TOKEN, apps={"directory": app_dir}, cli_parse_args=False)
-        config.set_validated_app_manifests()
-        result = config.apps.manifests
-
-        # Should auto-detect the app
-        assert "apps.test_app.TestApp" in result, "Expected to find 'apps.test_app.TestApp' in detected apps"
-        manifest = result["apps.test_app.TestApp"]
-        assert manifest.filename == "test_app.py", f"Expected filename to be 'test_app.py', got {manifest.filename}"
-        assert manifest.class_name == "TestApp", f"Expected class_name to be 'TestApp', got {manifest.class_name}"
-        assert manifest.enabled is True, f"Expected enabled to be True, got {manifest.enabled}"
-
-    def test_hassette_config_autodetect_can_be_disabled(self, tmp_path: Path):
-        """Test that autodetect_apps can be disabled in HassetteConfig."""
-        # Create a temporary app directory with an app
-        app_dir = tmp_path / "apps"
-        app_dir.mkdir()
-
-        app_file = app_dir / "test_app.py"
-        app_file.write_text(
-            dedent("""
-            from hassette import App, AppConfig
-
-            class TestApp(App[AppConfig]): ...
-        """)
-        )
-
-        # Create config with auto-detect disabled
-        config = HassetteConfig(
-            token=TEST_TOKEN,
-            apps={"directory": app_dir, "autodetect": False},
-            cli_parse_args=False,
-        )
-
-        # Should not auto-detect any apps (no manual apps configured either)
-        assert len(config.apps.apps) == 0, f"Expected 0 apps, got {len(config.apps.apps)}"
-
-    @pytest.mark.parametrize("ext", [".py", ""])
-    def test_defined_filename_without_extension_is_handled(self, tmp_path: Path, ext: str):
-        """If we define something in hassette.toml but forget the .py extension, we shouldn't load it twice.
-
-        We handle the missing .py in the AppManifest, but we need to make sure that the auto-detect
-        logic also handles this case correctly.
-
-        """
-        # Create a temporary app directory with an app
-        app_dir = tmp_path / "apps"
-        app_dir.mkdir()
-
-        app_file = app_dir / "priority_app.py"
-        app_file.write_text(
-            dedent("""
-            from hassette import App, AppConfig
-
-            class MyConfig(AppConfig):
-                custom: str = "auto-app"
-
-            class AutoDetectedApp(App[AppConfig]): ...
-        """)
-        )
-
-        # Create config with manual app configuration that conflicts
-        config = HassetteConfig(
-            token=TEST_TOKEN,
-            apps={
-                "directory": app_dir,
-                "apps": {
-                    "AutoDetectedApp": {
-                        "filename": f"priority_app{ext}",
-                        "class_name": "AutoDetectedApp",
-                        "enabled": True,  # Different from auto-detect default
-                        "config": {"custom": "value"},
-                    }
-                },
-            },
-            cli_parse_args=False,
-        )
-
-        config.set_validated_app_manifests()
-        result = config.apps.manifests
-
-        # Should have the manual configuration, not auto-detected
-        assert len(result) == 1, f"Expected 1 app, got {len(result)}"
-        assert "AutoDetectedApp" in result, "Expected to find 'priority_app.AutoDetectedApp' in detected apps"
-        manifest = result["AutoDetectedApp"]
-        assert manifest.enabled is True, f"Expected enabled to be True, got {manifest.enabled}"
-        assert manifest.app_config[0]["custom"] == "value", (
-            f"Expected custom config to be 'value', got {manifest.app_config[0]['custom']}"
-        )
-
-    def test_hassette_config_manual_apps_take_precedence(self, tmp_path: Path):
-        """Test that manually configured apps take precedence over auto-detected ones."""
-        # Create a temporary app directory with an app
-        app_dir = tmp_path / "apps"
-        app_dir.mkdir()
-
-        app_file = app_dir / "priority_app.py"
-        app_file.write_text(
-            dedent("""
-            from hassette import App, AppConfig
-
-            class AutoDetectedApp(App[AppConfig]): ...
-        """)
-        )
-
-        # Create config with manual app configuration that conflicts
-        config = HassetteConfig(
-            token=TEST_TOKEN,
-            apps={
-                "directory": app_dir,
-                "apps": {
-                    "priority_app.AutoDetectedApp": {
-                        "filename": "priority_app.py",
-                        "class_name": "AutoDetectedApp",
-                        "enabled": False,  # Different from auto-detect default
-                        "config": {"custom": "value"},
-                    }
-                },
-            },
-            cli_parse_args=False,
-        )
-
-        config.set_validated_app_manifests()
-        result = config.apps.manifests
-
-        # Should have the manual configuration, not auto-detected
-        assert len(result) == 1, f"Expected 1 app, got {len(result)}"
-        assert "priority_app.AutoDetectedApp" in result, (
-            "Expected to find 'priority_app.AutoDetectedApp' in detected apps"
-        )
-        manifest = result["priority_app.AutoDetectedApp"]
-        assert manifest.enabled is False, f"Expected enabled to be False, got {manifest.enabled}"
-        assert manifest.app_config[0]["custom"] == "value", (
-            f"Expected custom config to be 'value', got {manifest.app_config[0]['custom']}"
-        )
-
-    def test_hassette_config_combines_manual_and_autodetected(self, tmp_path: Path):
-        """Test that HassetteConfig combines manual and auto-detected apps correctly."""
-        # Create a temporary app directory with multiple apps
-        app_dir = tmp_path / "apps"
-        app_dir.mkdir()
-
-        # Auto-detected app
-        auto_file = app_dir / "auto_app.py"
-        auto_file.write_text(
-            dedent("""
-            from hassette import App, AppConfig
-
-            class AutoApp(App[AppConfig]): ...
-        """)
-        )
-
-        # Manually configured app
-        manual_file = app_dir / "manual_app.py"
-        manual_file.write_text(
-            dedent("""
-            from hassette import App, AppConfig
-
-            class ManualApp(App[AppConfig]): ...
-        """)
-        )
-
-        # Create config with one manual app
-        config = HassetteConfig(
-            token=TEST_TOKEN,
-            apps={
-                "directory": app_dir,
-                "apps": {
-                    "manual_app": {
-                        "filename": "manual_app.py",
-                        "class_name": "ManualApp",
-                        "config": {"manual": True},
-                    }
-                },
-            },
-            cli_parse_args=False,
-        )
-
-        config.set_validated_app_manifests()
-        result = config.apps.manifests
-
-        # Should have both manual and auto-detected apps
-        assert len(result) == 2, f"Expected 2 apps, got {len(result)}"
-        assert "manual_app" in result, "Expected to find 'manual_app' in detected apps"
-        assert "apps.auto_app.AutoApp" in result, "Expected to find 'apps.auto_app.AutoApp' in detected apps"
-
-        # Manual app should preserve config
-        manual_manifest = result["manual_app"]
-        assert manual_manifest.app_config[0]["manual"] is True, (
-            f"Expected manual config to be True, got {manual_manifest.app_config[0]['manual']}"
-        )
-
-        # Auto-detected app should have default config
-        auto_manifest = result["apps.auto_app.AutoApp"]
-        assert auto_manifest.enabled is True, f"Expected enabled to be True, got {auto_manifest.enabled}"
