+++ tests/unit/web/test_auth.py
@@ -56,7 +56,7 @@ def _propagate_hassette_logger() -> None:
 
     Some other test in the session may have left ``propagate`` set to False (e.g. via
     ``enable_basic_logging()``); caplog relies on propagation to the root logger. Same
-    workaround as ``tests/unit/test_autodetect_apps.py``.
+    workaround as ``tests/unit/test_validate_apps.py``.
     """
     logging.getLogger("hassette").propagate = True
